requirements.txt
lark-parser
PyAutoGUI
opencv-python
SQLAlchemy==2.0.0b2
dill

pandas
seaborn
